Fix warnings from previous commit

+++ lib/Module/ReturnToVoidFunctionPass.cpp
@@ -231,7 +231,7 @@ int klee::ReturnToVoidFunctionPass::replaceCall(CallInst *origCallInst, Function
   {
     // klee_error("Wrapper has bad signature! LLVM refuses to create call.");
     klee_warning("\e[1;35mWrapper has bad signature! LLVM refuses to create call. Args.size()=%d, FTy->getNumParams()=%d, FTY->isVarArg()=%d",
-      Args.size(), FTy->getNumParams(), (FTy->isVarArg()));
+      (int)Args.size(), FTy->getNumParams(), (FTy->isVarArg()));
     return 1;
   }
   for (unsigned i = 0; i != Args.size(); ++i)
